Add information about Plugin Dependency in plugin cards
Edit the dependencies property to be an array of object in which the key is the function to check and the value is the Plugin checked.
Add the plugin card bottom div when dependencies are unsatisfied for a plugin within the Entrepôt tab of the "Add Plugins" administration screen.
Add a unit test for the plugin dependency check function.

See #4