Add restrictions to the WordPress Plugin code editor
Plugin core files should not be edited for many reasons starting with changes lost on plugin updates. This one is very dangerous as some users might not apply security updates to avoid losing their changes. That is why Entrepôt is using a specific Plugin Header Tag to allow the feature when it is interesting. For instance the wp-idea-stream-custom.php file or the bp-custom.php file WP Idea Stream or BuddyPress are using to allow users to edit the plugin behavior. For this cases, simply add a "Allow File Edits" Header Tag set to True and you will be able to edit the plugins/custom function files