Predict the unpredictable :)
As Gutenberg is *arbitrary* deregistering the `wp-api-request` script to replace it with a new implementation, we cannot trust anymore WordPress Core way of setting the wpApiSettings. As a result to be sure to be able to get these settings for this plugin need, we have to add it to our script in case Gutenberg is active. IMHO, it is a bit embarassing Gutenberg took this decision as a Plugin. This kind of change would have probably need some explanations on the make.w.org... but Gutenberg is not a plugin like all the others ;)

See https://github.com/WordPress/gutenberg/pull/7329